Cleaning up and moving modules
