+++ CeresBbModules/ceres_drs/CeresBbDynamicTableModule.php
@@ -12,7 +12,7 @@ class DynamicTable extends FLBuilderModule {
         parent::__construct(array(
             'name'            => __( 'CERES DynamicTable BB module', 'fl-builder' ),
             'description'     => __( 'Just Testing', 'fl-builder' ),
-            'group'           => __( 'CERES Classic', 'fl-builder' ),
+            'group'           => __( 'CERES Legacy', 'fl-builder' ),
             'category'        => __( 'CERES', 'fl-builder' ),
             'dir'             => CERES_BB_MODULES_DIR . 'CeresBbModules/ceres_drs',
             'url'             => CERES_BB_MODULES_URL . 'CeresBbModules/ceres_drs',
