+++ CeresBbModules/ceres_image/includes/frontend.php
@@ -1,3 +1,3 @@
 <div class="fl-example-text" id="textarea-field-<?php echo $module->node; ?>">
-    <?php echo $module->render(); ?>
+<!--    --><?php //echo $module->render(); ?>
 </div>
