--- CeresBbModules/ceres_image/includes/input-form.css
@@ -1,8 +1,0 @@
-/* styles.css */
-.img-container .zoom-image {
-    transition: transform 0.5s ease;
-}
-
-.img-container:hover .zoom-image {
-    transform: scale(1.2);
-}
