--- CeresBbModules/ceres_simple/includes/frontend.php
@@ -1,11 +1,0 @@
-
-<div class="fl-example-text">
-<?php
-// var_dump($settings);
-// die();
-?>
-    <?php echo $settings->text; ?>
-    <?php echo $module->test(); ?>
-</div>
-
-
