--- CeresBbModules/ceres_table/includes/frontend.php
@@ -1,3 +1,0 @@
-<div class="fl-example-text" id="textarea-field-<?php echo $module->node; ?>">
-    <?php echo $module->test(); ?>
-</div>
