--- CeresBbModules/ceres_tmodule/includes/frontend.js.php
@@ -1,2 +1,0 @@
-console.log('Text: <?php echo $settings->textarea_field; ?>');
-console.log('Text: ' + document.getElementById('textarea-field-<?php echo $module->node; ?>').innerHTML);
