--- CeresBbModules/ceres_tmodule/includes/frontend.php
@@ -1,4 +1,0 @@
-<div class="fl-example-text" id="textarea-field-<?php echo $module->node; ?>">
-<!--    --><?php //$module->render(); ?>
-    <?php echo $settings->textarea_field; ?>
-</div>
