--- CeresBbModules/ceres_tmodule/includes/input-form.php
@@ -1,6 +1,0 @@
-<div class="fl-example-text">
-    <form method="post">
-        <input type="text" name="textarea_field_<?php echo $module_id; ?>" id="textarea_field_<?php echo $module_id; ?>" placeholder="Enter your text here">
-        <input type="submit" name="button2_<?php echo $module_id; ?>" value="Button2"/>
-    </form>
-</div>
